feat: (7/N) refactor ServiceWorker to test setLevel improvements
